[lsp: document synchronisation]
Thus far lsp requests were embraced by lsp document open and close.
Beneath the additional number of communications with the server this
also caused the completion to report different results. It seems that
newer clangd implementations are not quick enough to return proper
completions if the lsp completion request is sent immediately after the
lsp document open. However, if the completion request is sent a while
after the document open notification has been sent, the resulting
completions are useful.
To achieve this, instead of opening and closing a document for each lsp
request, open and close are linked to vim buffer events. To communicate
buffer updates to the language server becomplete#lsp#document#update()
has been added.

To allow sending the document open notification during buffer
initialisation, it is required to get the server initialised beforehand.
Hence, the language server startup has been moved to the plugin main
file.