[ctags: fix server indexing]
This change updates the ctags indexing call to only index files of the
given type. During server initialisation the directory of the current
buffer is used as root, instead of the current directory.

Thus far, if be-complete used the ctags implementation for any file
type, all files of types supported by the ctags plugin layer were
indexed. For instance, if a Makefile is indexed also all c files would,
even if a language server would be configured for c files already.
This caused increased startup times and double indexing.