[completion: in strings and comments]
Thus far completion within comments and string depends on the language
server, some might support it, clangd for instance doesn't. To still
allow for some completion within comments and strings, this change adds
a check to determine of the cursor is within a corresponding syntax
region, if so, vim's keyword completion is triggered.