[completion: function signature]
Add a special case for completions triggered on a "(" character. Thus
far this didn't result in any completion, since the word under the
cursor is empty, which explicitly aborts completion.
With this change, completion is triggered for the word in front of the
"(" character, potentially restarting the function and thus signature
completion.