[fix: function argument selection]
issue:

During function argument selection, if the line doesn't end with the
last argument marker, the forward selection of arguments stops on the
last argument instead of continuing with the first.

reason:

Forward argument selection checks if the cursor is at the end of the
line. If it isn't, it assumes another argument to follow. If this
assumption is not true nothing is selected.

fix:

As for backward argument selection, check if a left-side argument marker
has been found, if not restart searching at the beginning of the line.