[fix: language server document update]
issue:

Changes in a header file are visible in the completion result of a
related source file.

reason:

File changes are only communicated to the language server before calling
a language server feature, such as completion. Until a language server
feature is called, writes to a file only cause a 'modified' flag to be
set.

As a consequence, when a header file is written its 'modified' flag is
set without communicating the updated content to the language server.
Hence, the update in the header file is not considered when calling for
completion in the source file.

fix:

Remove the 'doc_update' server callback and instead of just setting the
'modified' flag for a file, immediately communicating the update to the
server.