[fix: language server document update]
issue:

Deletions of e.g. variable declarations and struct members are not
visible in completion results for other source files, even if the header
file had been saved already.

reason:

Apparently the "textDocument/didSave" notification is expected by
language servers that have the "textDocumentSync.save" capability set,
as clangd does.

fix:

Add the "doc_write" be-complete server callback to signal buffer writes
to servers. This callback is implemented for lsp and a nop for ctags.