[fix: key not present for symlinked files]
issue:

Requesting any be-complete functionality, like completion or symbol
information, triggered a "key not present in dictionary" error for some
files.

reason:

The problematic files contain a symlink which is resolved different for
expand("<afile>") and expand("%"), with "<afile>" being used within
autocommands and "%" anywhere else. Hence, the file is registered to the
language server with a different path than the one used for later
requests.

fix:

Resolve symlinks for expand("<afile>") be wrapping it into resolve().