[fix: completion for gopls]
issue:

Completion doesn't work for the go language server gopls.

reason:

The response to the complete request of gopls doesn't provide a value
for 'insertText', which is used by be-complete to identify the text for
a completion item. This is according to the lsp specification which
states to ignore 'insertText' if 'textEdit' is provided.

fix:

Use 'textEdit' as the main source to identify the text for a completion
item and 'insertText' as fallback, if 'textEdit' is missing.