[lsp: fix completion]
issue:

vim throws error E1098 in the lsp completion item filter function, due
to using something not iterable for a loop.

reason:

The supposed item list, containing the completion results and supposed
to be used as the iterable for the loop, is a dictionary, instead of a
list.

This case occurs if the language server returns a list of
CompletiionItems (an actual list) instead of a CompletionList (a dict),
cf. language server protocol specification.
If CompletionList is given, the filter function extracts the "items" and
proceeds to parse them. In the case of a list of CompletiionItems, the
default case of the vim get function was intended to trigger, returning
the list as is. However, the vim get function acts differently for dicts
and list. Hence, instead of triggering the default case, since "items"
is not within the list, the first item of the list is returned. Since
this item is a dict, vim raises E1098 when trying to iterate over it.

fix:

Check the type of the lsp response rather than relying on the get
function's default argument.